+++ getresponse-official/core/class-gr-configuration-validator-exception.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GR\WordPress\Core;
+
+use Exception;
+use WP_Error;
+
+class Gr_Configuration_Validator_Exception extends Exception {
+}