+++ getresponse-official/core/class-gr-rest-api-service.php
@@ -6,18 +6,20 @@ namespace GR\WordPress\Core;
 
 class Gr_Rest_Api_Service {
 
+
 	public const GR_WEB_CONNECT_SNIPPET        = 'gr4wp-web-connect-snippet';
-	public const GR_RECOMMENDATION_SNIPPET     = 'gr4wp-recommendation-snippet';
 	public const GR_LIVE_SYNC_URL              = 'gr4wp-live-sync-url';
 	public const GR_LIVE_SYNC_TYPE             = 'gr4wp-live-sync-type';
 	public const GR_MARKETING_CONSENT_TEXT     = 'gr4wp-marketing-consent-text';
 	public const GR_SHOP_ID                    = 'gr4wp-shop-id';
 	public const INTEGRATE_WITH_CONTACT_FORM_7 = 'gr4wp_integrate_with_contact_form_7';
 
+	public function __construct() {
+	}
+
 	public function get_configuration(): Gr_Configuration {
 		return new Gr_Configuration(
 			(string) $this->get_gr_option( self::GR_WEB_CONNECT_SNIPPET ),
-			(string) $this->get_gr_option( self::GR_RECOMMENDATION_SNIPPET ),
 			(string) $this->get_gr_option( self::GR_LIVE_SYNC_URL ),
 			(string) $this->get_gr_option( self::GR_LIVE_SYNC_TYPE ),
 			(string) $this->get_gr_option( self::GR_MARKETING_CONSENT_TEXT ),
@@ -28,7 +30,6 @@ class Gr_Rest_Api_Service {
 
 	public function update_configuration( Gr_Configuration $configuration ): void {
 		$this->update_gr_option( self::GR_WEB_CONNECT_SNIPPET, $configuration->get_web_connect_snippet() );
-		$this->update_gr_option( self::GR_RECOMMENDATION_SNIPPET, $configuration->get_recommendation_snippet() );
 		$this->update_gr_option( self::GR_LIVE_SYNC_URL, $configuration->get_live_sync_url() );
 		$this->update_gr_option( self::GR_LIVE_SYNC_TYPE, $configuration->get_live_sync_type() );
 		$this->update_gr_option( self::GR_MARKETING_CONSENT_TEXT, $configuration->get_marketing_consent_text() );
@@ -38,7 +39,6 @@ class Gr_Rest_Api_Service {
 
 	public function clear_configuration(): void {
 		$this->update_gr_option( self::GR_WEB_CONNECT_SNIPPET, '' );
-		$this->update_gr_option( self::GR_RECOMMENDATION_SNIPPET, '' );
 		$this->update_gr_option( self::GR_LIVE_SYNC_URL, '' );
 		$this->update_gr_option( self::GR_LIVE_SYNC_TYPE, '' );
 		$this->update_gr_option( self::GR_MARKETING_CONSENT_TEXT, '' );
@@ -48,7 +48,6 @@ class Gr_Rest_Api_Service {
 
 	public function delete_configuration(): void {
 		$this->delete_gr_option( self::GR_WEB_CONNECT_SNIPPET );
-		$this->delete_gr_option( self::GR_RECOMMENDATION_SNIPPET );
 		$this->delete_gr_option( self::GR_LIVE_SYNC_URL );
 		$this->delete_gr_option( self::GR_LIVE_SYNC_TYPE );
 		$this->delete_gr_option( self::GR_MARKETING_CONSENT_TEXT );